feat: refactor domain and infrastructure layers with value objects and enhanced exception handling
