+++ fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/domain/model/order/OrderItem.java
@@ -1,0 +1,18 @@
+package cat.itacademy.fruitorderapimongo.domain.model.order;
+
+import cat.itacademy.fruitorderapimongo.domain.model.Fruit;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+public class OrderItem {
+    @Setter
+    private Fruit fruit;
+    @Setter
+    private Double quantityInKg;
+
+    public OrderItem(Fruit fruit, Double quantityInKg) {
+        this.fruit = fruit;
+        this.quantityInKg = quantityInKg;
+    }
+}
