+++ fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/domain/model/valueobject/DataOrder.java
@@ -1,0 +1,25 @@
+package cat.itacademy.fruitorderapimongo.domain.model.valueobject;
+
+import cat.itacademy.fruitorderapimongo.infrastructure.exception.InvalidDateException;
+import lombok.Getter;
+
+import java.time.LocalDate;
+
+@Getter
+public class DataOrder {
+    private final LocalDate deliveryDate;
+
+    private DataOrder(LocalDate deliveryDate) {
+        this.deliveryDate = deliveryDate;
+    }
+
+    public static DataOrder of(LocalDate deliveryDate) {
+        if (deliveryDate == null) {
+            throw new InvalidDateException("Date cannot be null");
+        }
+        if (!deliveryDate.isAfter(LocalDate.now())) {
+            throw new InvalidDateException("Date must be at least tomorrow");
+        }
+        return new DataOrder(deliveryDate);
+    }
+}
