+++ fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/domain/model/valueobject/Name.java
@@ -1,0 +1,24 @@
+package cat.itacademy.fruitorderapimongo.domain.model.valueobject;
+
+import cat.itacademy.fruitorderapimongo.infrastructure.exception.InvalidNameException;
+import lombok.Getter;
+
+@Getter
+public class Name {
+
+    private final String value;
+
+    private Name(String value) {
+        this.value = value;
+    }
+
+    public static Name of(String value) {
+        if (value == null) {
+            throw new InvalidNameException("Name cannot be null");
+        }
+        if (value.isBlank()) {
+            throw new InvalidNameException("Name cannot be blank");
+        }
+        return new Name(value);
+    }
+}
