--- fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/dto/fruit/FruitDtoResponse.java
@@ -1,6 +1,0 @@
-package cat.itacademy.fruitorderapimongo.dto.fruit;
-
-import cat.itacademy.fruitorderapimongo.dto.supplier.SupplierDtoResponse;
-
-public record FruitDtoResponse (String id, String name, Double weightKg, SupplierDtoResponse supplierDtoResponse) {
-}
