--- fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/dto/supplier/SupplierDtoResponse.java
@@ -1,4 +1,0 @@
-package cat.itacademy.fruitorderapimongo.dto.supplier;
-
-public record SupplierDtoResponse(String id, String name, String country) {
-}
