--- fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/exception/GlobalExceptionHandler.java
@@ -1,67 +1,0 @@
-package cat.itacademy.fruitorderapimongo.exception;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.http.converter.HttpMessageNotReadableException;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.time.LocalDateTime;
-import java.util.HashMap;
-import java.util.Map;
-
-@Slf4j
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<ErrorDetails> handleValidationErrors(MethodArgumentNotValidException ex) {
-        Map<String, String> errors = new HashMap<>();
-        ex.getBindingResult().getFieldErrors().forEach(error -> errors.put(error.getField(), error.getDefaultMessage()));
-        ErrorDetails details = new ErrorDetails(LocalDateTime.now(), HttpStatus.BAD_REQUEST.value(), "Validation failed", errors);
-        return new ResponseEntity<>(details, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(HttpMessageNotReadableException.class)
-    public ResponseEntity<ErrorDetails> handleNotReadable(HttpMessageNotReadableException ex) {
-        return buildResponse("Invalid request format or field types", HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(ResourceNotFoundException.class)
-    public ResponseEntity<ErrorDetails> handleNotFound(ResourceNotFoundException ex) {
-        return buildResponse(ex.getMessage(), HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(ResourceAlreadyExistsException.class)
-    public ResponseEntity<ErrorDetails> handleAlreadyExist(ResourceAlreadyExistsException ex) {
-        return buildResponse(ex.getMessage(), HttpStatus.CONFLICT);
-    }
-
-    @ExceptionHandler(ResourceHasDependenciesException.class)
-    public ResponseEntity<ErrorDetails> handleResourceHasDependencies(ResourceHasDependenciesException ex) {
-        return buildResponse(ex.getMessage(), HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(InvalidDateException.class)
-    public ResponseEntity<ErrorDetails> handleInvalidDate(InvalidDateException ex) {
-        return buildResponse(ex.getMessage(), HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ErrorDetails> handleGeneralError(Exception ex) {
-        log.error("Unexpected error: ", ex);
-        return buildResponse("Internal Server Error", HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-
-    private ResponseEntity<ErrorDetails> buildResponse(String message, HttpStatus status) {
-        return new ResponseEntity<>(new ErrorDetails(LocalDateTime.now(), status.value(), message), status);
-    }
-
-    public record ErrorDetails(LocalDateTime timestamp, int status, String message, Map<String, String> errors) {
-        public ErrorDetails(LocalDateTime timestamp, int status, String message) {
-            this(timestamp, status, message, null);
-        }
-    }
-}
