--- fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/exception/InvalidDateException.java
@@ -1,7 +1,0 @@
-package cat.itacademy.fruitorderapimongo.exception;
-
-public class InvalidDateException extends RuntimeException {
-    public InvalidDateException(String message) {
-        super(message);
-    }
-}
