--- fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/exception/ResourceAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package cat.itacademy.fruitorderapimongo.exception;
-
-public class ResourceAlreadyExistsException extends RuntimeException {
-    public ResourceAlreadyExistsException(String resourceName, String fieldName ) {
-        super(String.format("%s with name %s already exists", resourceName, fieldName));
-    }
-}
