--- fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/exception/ResourceHasDependenciesException.java
@@ -1,7 +1,0 @@
-package cat.itacademy.fruitorderapimongo.exception;
-
-public class ResourceHasDependenciesException extends RuntimeException {
-    public ResourceHasDependenciesException(String nameResource, String id) {
-        super(String.format("%s with id: %s has fruits associated", nameResource, id));
-    }
-}
