+++ fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/infrastructure/adapter/in/rest/OrderController.java
@@ -1,0 +1,30 @@
+package cat.itacademy.fruitorderapimongo.infrastructure.adapter.in.rest;
+
+import cat.itacademy.fruitorderapimongo.infrastructure.dto.order.OrderDtoRequest;
+import cat.itacademy.fruitorderapimongo.infrastructure.dto.order.OrderDtoResponse;
+import cat.itacademy.fruitorderapimongo.service.OrderService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+
+import java.net.URI;
+
+@RestController
+@RequestMapping ("/orders")
+public class OrderController {
+    private final OrderService orderService;
+
+    public OrderController(OrderService orderService) {
+        this.orderService = orderService;
+    }
+
+    @PostMapping
+    ResponseEntity<OrderDtoResponse> createOrder(@RequestBody OrderDtoRequest order){
+        OrderDtoResponse createdOrder = orderService.createOrder(order);
+        URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}").buildAndExpand(createdOrder.id()).toUri();
+        return ResponseEntity.created(location).body(createdOrder);
+    }
+}
