+++ fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/infrastructure/dto/fruit/FruitDtoResponse.java
@@ -1,0 +1,6 @@
+package cat.itacademy.fruitorderapimongo.infrastructure.dto.fruit;
+
+import cat.itacademy.fruitorderapimongo.infrastructure.dto.supplier.SupplierDtoResponse;
+
+public record FruitDtoResponse (String id, String name, Double weightKg, SupplierDtoResponse supplierDtoResponse) {
+}
