+++ fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/infrastructure/dto/order/OrderDtoResponse.java
@@ -1,0 +1,12 @@
+package cat.itacademy.fruitorderapimongo.infrastructure.dto.order;
+
+import java.time.LocalDate;
+import java.util.List;
+
+public record OrderDtoResponse(
+        String id,
+        String clientName,
+        LocalDate deliveryDate,
+        List<OrderItemDtoResponse> orderItems
+) {
+}
