+++ fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/infrastructure/dto/order/OrderItemDtoRequest.java
@@ -1,0 +1,15 @@
+package cat.itacademy.fruitorderapimongo.infrastructure.dto.order;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
+
+public record OrderItemDtoRequest(
+        @NotBlank(message = "fruitId cannot be blank")
+        String fruitId,
+
+        @NotNull(message = "quantityInKg cannot be empty")
+        @Positive(message = "quantityInKg must be positive")
+        Double quantityInKg
+) {
+}
