+++ fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/infrastructure/dto/order/OrderItemDtoResponse.java
@@ -1,0 +1,9 @@
+package cat.itacademy.fruitorderapimongo.infrastructure.dto.order;
+
+import cat.itacademy.fruitorderapimongo.infrastructure.dto.fruit.FruitDtoResponse;
+
+public record OrderItemDtoResponse(
+        FruitDtoResponse fruit,
+        Double quantityInKg
+) {
+}
