+++ fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/infrastructure/dto/supplier/SupplierDtoResponse.java
@@ -1,0 +1,4 @@
+package cat.itacademy.fruitorderapimongo.infrastructure.dto.supplier;
+
+public record SupplierDtoResponse(String id, String name, String country) {
+}
