+++ fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/infrastructure/exception/InvalidDateException.java
@@ -1,0 +1,7 @@
+package cat.itacademy.fruitorderapimongo.infrastructure.exception;
+
+public class InvalidDateException extends RuntimeException {
+    public InvalidDateException(String message) {
+        super(message);
+    }
+}
