+++ fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/infrastructure/exception/InvalidNameException.java
@@ -1,0 +1,7 @@
+package cat.itacademy.fruitorderapimongo.infrastructure.exception;
+
+public class InvalidNameException extends RuntimeException {
+    public InvalidNameException(String message) {
+        super(message);
+    }
+}
