+++ fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/infrastructure/exception/InvalidPriceException.java
@@ -1,0 +1,7 @@
+package cat.itacademy.fruitorderapimongo.infrastructure.exception;
+
+public class InvalidPriceException extends RuntimeException {
+    public InvalidPriceException(String message) {
+        super(message);
+    }
+}
