+++ fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/infrastructure/exception/ResourceHasDependenciesException.java
@@ -1,0 +1,7 @@
+package cat.itacademy.fruitorderapimongo.infrastructure.exception;
+
+public class ResourceHasDependenciesException extends RuntimeException {
+    public ResourceHasDependenciesException(String nameResource, String id) {
+        super(String.format("%s with id: %s has fruits associated", nameResource, id));
+    }
+}
