+++ fruit-order-api-mongo/src/main/java/cat/itacademy/fruitorderapimongo/infrastructure/mapper/SupplierMapper.java
@@ -1,0 +1,22 @@
+package cat.itacademy.fruitorderapimongo.infrastructure.mapper;
+
+import cat.itacademy.fruitorderapimongo.infrastructure.dto.supplier.SupplierDtoRequest;
+import cat.itacademy.fruitorderapimongo.infrastructure.dto.supplier.SupplierDtoResponse;
+import cat.itacademy.fruitorderapimongo.domain.model.Supplier;
+import cat.itacademy.fruitorderapimongo.domain.model.valueobject.Country;
+import cat.itacademy.fruitorderapimongo.domain.model.valueobject.Name;
+
+
+public class SupplierMapper {
+    public static Supplier toEntity(SupplierDtoRequest supplierDtoRequest) {
+        return new Supplier(Name.of(supplierDtoRequest.name()), Country.of(supplierDtoRequest.country()));
+    }
+
+    public static SupplierDtoResponse toDto(Supplier supplier) {
+        return new SupplierDtoResponse(
+                supplier.getId(),
+                supplier.getName().getValue(),
+                supplier.getCountry().getValue()
+        );
+    }
+}
